Fix CMemPool method params using wrong type in implementation

+++ source_sdk/tier1/MemPool.cpp
@@ -219,7 +219,7 @@ void* CMemoryPool::AllocZero()
 // Purpose: Allocs a single block of memory from the pool.  
 // Input  : amount - 
 //-----------------------------------------------------------------------------
-void *CMemoryPool::Alloc( unsigned int amount )
+void *CMemoryPool::Alloc( size_t amount )
 {
 	void *returnBlock;
 
@@ -260,7 +260,7 @@ void *CMemoryPool::Alloc( unsigned int amount )
 // Purpose: Allocs a single block of memory from the pool, zeroes the memory before returning
 // Input  : amount - 
 //-----------------------------------------------------------------------------
-void *CMemoryPool::AllocZero( unsigned int amount )
+void *CMemoryPool::AllocZero( size_t amount )
 {
 	void *mem = Alloc( amount );
 	if ( mem )
